Begin ripping out code for unsupported Celeste versions
