+++ Celeste.Mod.mm/Mod/Core/CoreModule.cs
@@ -137,7 +137,7 @@ namespace Celeste.Mod.Core {
                     ((Engine.Scene as Overworld)?.IsCurrent<OuiTitleScreen>() ?? false) ||
                     (Engine.Scene is GameLoader)
                 ,
-                Button = (patch_VirtualButton_InputV2) Input.MenuConfirm
+                Button = (patch_VirtualButton) Input.MenuConfirm
             };
         }
