+++ Celeste.Mod.mm/Mod/Everest/ButtonBinding.cs
@@ -21,94 +21,31 @@ namespace Celeste.Mod {
             set => SetButtons(value);
         }
 
-        [MonoModIgnore]
-        private extern List<Buttons> GetButtons();
 
-        [MonoModIgnore]
-        private extern void SetButtons(List<Buttons> value);
-
-        [MonoModIfFlag("V1:Input")]
-        private List<Buttons> _Buttons;
-
-        [MonoModIfFlag("V1:Input")]
-        [MonoModPatch("GetButtons")]
         [MonoModReplace]
-        private List<Buttons> GetButtonsV1() {
-            return _Buttons;
-        }
+        private List<Buttons> GetButtons() => Binding.Controller;
 
-        [MonoModIfFlag("V1:Input")]
-        [MonoModPatch("SetButtons")]
-        [MonoModReplace]
-        private void SetButtonsV1(List<Buttons> value) {
-            _Buttons = value;
-        }
 
-        [MonoModIfFlag("V2:Input")]
-        [MonoModPatch("GetButtons")]
         [MonoModReplace]
-        private List<Buttons> GetButtonsV2() {
-            return Binding.Controller;
-        }
-
-        [MonoModIfFlag("V2:Input")]
-        [MonoModPatch("SetButtons")]
-        [MonoModReplace]
-        private void SetButtonsV2(List<Buttons> value) {
-            Binding.Controller = value;
-        }
+        private void SetButtons(List<Buttons> value) => Binding.Controller = value;
 
         public List<Keys> Keys {
             get => GetKeys();
             set => SetKeys(value);
         }
 
-        [MonoModIgnore]
-        private extern List<Keys> GetKeys();
 
-        [MonoModIgnore]
-        private extern void SetKeys(List<Keys> value);
-
-        [MonoModIfFlag("V1:Input")]
-        private List<Keys> _Keys;
-
-        [MonoModIfFlag("V1:Input")]
-        [MonoModPatch("GetKeys")]
         [MonoModReplace]
-        private List<Keys> GetKeysV1() {
-            return _Keys;
-        }
+        private List<Keys> GetKeys() => Binding.Keyboard;
 
-        [MonoModIfFlag("V1:Input")]
-        [MonoModPatch("SetKeys")]
         [MonoModReplace]
-        private void SetKeysV1(List<Keys> value) {
-            _Keys = value;
-        }
+        private void SetKeys(List<Keys> value) => Binding.Keyboard = value;
 
-        [MonoModIfFlag("V2:Input")]
-        [MonoModPatch("GetKeys")]
-        [MonoModReplace]
-        private List<Keys> GetKeysV2() {
-            return Binding.Keyboard;
-        }
-
-        [MonoModIfFlag("V2:Input")]
-        [MonoModPatch("SetKeys")]
-        [MonoModReplace]
-        private void SetKeysV2(List<Keys> value) {
-            Binding.Keyboard = value;
-        }
-
-        [MonoModIfFlag("V2:Input")]
         private Binding _Binding;
 
-        [MonoModIfFlag("V2:Input")]
         [YamlIgnore] // Binding uses XmlIgnores and migrating old mappings is hard.
         public Binding Binding {
-            [MonoModIfFlag("V2:Input")]
             get => _Binding;
-            [MonoModIfFlag("V2:Input")]
             private set => _Binding = value;
         }
 
@@ -124,8 +61,8 @@ namespace Celeste.Mod {
         public bool Repeating => Button?.Repeating ?? false;
 
         [YamlIgnore]
-        [Obsolete("Please move to Binding ASAP (ideally after the new Celeste input system gets out of beta).")]
-        public List<patch_VirtualButton_InputV1.Node> Nodes {
+        [Obsolete("Please move to Binding ASAP.")]
+        public List<patch_VirtualButton.Node> Nodes {
             get => Button?.Nodes;
             set {
                 //No null-coalescing assignment operator in C#7.3
@@ -154,7 +91,7 @@ namespace Celeste.Mod {
         #endregion
 
         [YamlIgnore]
-        public patch_VirtualButton_InputV1 Button;
+        public patch_VirtualButton Button;
 
         public ButtonBinding()
             : this(0) {
@@ -164,22 +101,8 @@ namespace Celeste.Mod {
             Init(buttons, keys);
         }
 
-        [MonoModIgnore]
-        private extern void Init(Buttons buttons, params Keys[] keys);
-
-        [MonoModIfFlag("V1:Input")]
-        [MonoModPatch("Init")]
         [MonoModReplace]
-        [Obsolete]
-        private void InitV1(Buttons buttons, params Keys[] keys) {
-            Buttons = Enum.GetValues(typeof(Buttons)).Cast<Buttons>().Where(b => (buttons & b) == b).ToList();
-            Keys = new List<Keys>(keys);
-        }
-
-        [MonoModIfFlag("V2:Input")]
-        [MonoModPatch("Init")]
-        [MonoModReplace]
-        private void InitV2(Buttons buttons, params Keys[] keys) {
+        private void Init(Buttons buttons, params Keys[] keys) {
             Binding = new Binding() {
                 Controller = Enum.GetValues(typeof(Buttons)).Cast<Buttons>().Where(b => (buttons & b) == b).ToList(),
                 Keyboard = new List<Keys>(keys)
