+++ Celeste.Mod.mm/Mod/Everest/Extensions.cs
@@ -307,19 +307,19 @@ namespace Celeste.Mod {
         /// </summary>
         /// <param name="vbtn">The virtual button to map the buttons to.</param>
         /// <param name="buttons">The buttons to map.</param>
-        public static void AddButtons(this patch_VirtualButton_InputV2 vbtn, List<Buttons> buttons) {
+        public static void AddButtons(this patch_VirtualButton vbtn, List<Buttons> buttons) {
             foreach (Buttons btn in buttons) {
                 if (btn == Buttons.LeftTrigger) {
-                    vbtn.Nodes.Add(new patch_VirtualButton_InputV2.PadLeftTrigger(Input.Gamepad, 0.25f));
+                    vbtn.Nodes.Add(new patch_VirtualButton.PadLeftTrigger(Input.Gamepad, 0.25f));
                     continue;
                 }
 
                 if (btn == Buttons.RightTrigger) {
-                    vbtn.Nodes.Add(new patch_VirtualButton_InputV2.PadRightTrigger(Input.Gamepad, 0.25f));
+                    vbtn.Nodes.Add(new patch_VirtualButton.PadRightTrigger(Input.Gamepad, 0.25f));
                     continue;
                 }
 
-                vbtn.Nodes.Add(new patch_VirtualButton_InputV2.PadButton(Input.Gamepad, btn));
+                vbtn.Nodes.Add(new patch_VirtualButton.PadButton(Input.Gamepad, btn));
             }
         }
