+++ Celeste.Mod.mm/Mod/UI/ModuleSettingsButtonConfigUI.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Input;
+﻿using Microsoft.Xna.Framework.Input;
 using Monocle;
 using MonoMod;
 using System;
@@ -7,9 +6,7 @@ using System.Collections.Generic;
 using System.Reflection;
 
 namespace Celeste.Mod {
-    // This MUST keep its old name. If V1 support gets dropped, rename V2 to this and use MonoModLinkFrom V2 -> this.
-    [Obsolete]
-    public class ModuleSettingsButtonConfigUI : patch_ButtonConfigUI_InputV1 {
+    public class ModuleSettingsButtonConfigUI : patch_ButtonConfigUI {
 
         public EverestModule Module;
 
@@ -23,18 +20,7 @@ namespace Celeste.Mod {
 
             Module = module;
             // Base already reloads too early before the module has been set.
-            Reload();
-        }
-
-        protected override string GetLabel(int mapping) {
-            return Bindings[mapping].Name;
-        }
-
-        protected override List<Buttons> GetRemapList(int remapping, Buttons newBtn) {
-            return Bindings[remapping].Binding.Buttons;
-        }
-
-        protected override void ForceRemapAll() {
+            Reload(2);
         }
 
         public override void Reload(int index = -1) {
@@ -43,7 +29,7 @@ namespace Celeste.Mod {
 
             Clear();
             Add(new Header(Dialog.Clean("BTN_CONFIG_TITLE")));
-            Add(new Info());
+            Add(new InputMappingInfo(true));
 
             Bindings.Clear();
 
@@ -79,8 +65,8 @@ namespace Celeste.Mod {
 
                     DefaultButtonBindingAttribute defaults = prop.GetCustomAttribute<DefaultButtonBindingAttribute>();
 
-                    Bindings.Add(new ButtonBindingEntry(name, binding, defaults));
-                    AddButtonConfigLine(mapping, defaults != null && defaults.Button != 0 && defaults.ForceDefaultButton ? ForceDefaultButton(defaults.Button, binding.Buttons) : binding.Buttons);
+                    Bindings.Add(new ButtonBindingEntry(binding, defaults));
+                    AddMapForceLabel(name, binding.Binding);
                 }
             }
 
@@ -88,85 +74,31 @@ namespace Celeste.Mod {
             Add(new Button(Dialog.Clean("KEY_CONFIG_RESET")) {
                 IncludeWidthInMeasurement = false,
                 AlwaysCenter = true,
-                OnPressed = () => {
-                    foreach (ButtonBindingEntry entry in Bindings) {
-                        entry.Binding.Buttons.Clear();
-                        if (entry.Defaults != null && entry.Defaults.Button != 0)
-                            entry.Binding.Buttons.Add(entry.Defaults.Button);
-                    }
-                    Input.Initialize();
-                    Reload(Selection);
-                }
+                OnPressed = () => ResetPressed()
             });
 
             if (index >= 0)
                 Selection = index;
         }
 
+        public override void Reset() {
+            foreach (ButtonBindingEntry entry in Bindings) {
+                entry.Binding.Binding.Controller.Clear();
+                if (entry.Defaults != null && entry.Defaults.Button != 0)
+                    entry.Binding.Binding.Controller.Add(entry.Defaults.Button);
+            }
+            Input.Initialize();
+            Reload(Selection);
+        }
+
         protected class ButtonBindingEntry {
 
-            public string Name;
             public ButtonBinding Binding;
             public DefaultButtonBindingAttribute Defaults;
 
-            public ButtonBindingEntry(string name, ButtonBinding binding, DefaultButtonBindingAttribute defaults) {
-                Name = name;
+            public ButtonBindingEntry(ButtonBinding binding, DefaultButtonBindingAttribute defaults) {
                 Binding = binding;
                 Defaults = defaults;
-            }
-
-        }
-
-        // This MUST exist here as it's missing in V2 yet this class gets patched into both V1 and V2 environments and aaaa~
-        [Obsolete]
-        public new class Info : Item {
-
-            private List<object> info = new List<object>();
-
-            public Info() {
-                string[] texts = Dialog.Clean("BTN_CONFIG_INFO", null).Split('|');
-                if (texts.Length == 3) {
-                    info.Add(texts[0]);
-                    info.Add(Input.MenuConfirm);
-                    info.Add(texts[1]);
-                    info.Add(Input.MenuJournal);
-                    info.Add(texts[2]);
-                }
-            }
-
-            public override float LeftWidth()
-                => 100f;
-
-            public override float Height()
-                => ActiveFont.LineHeight * 2f;
-
-            public override void Render(Vector2 position, bool highlighted) {
-                Color textColor = Color.Gray * Ease.CubeOut(Container.Alpha);
-                Color strokeColor = Color.Black * Ease.CubeOut(Container.Alpha);
-                Color btnColor = Color.White * Ease.CubeOut(Container.Alpha);
-
-                float taken = 0f;
-                for (int i = 0; i < info.Count; i++) {
-                    if (info[i] is string text) {
-                        taken += ActiveFont.Measure(text).X * 0.6f;
-
-                    } else if (info[i] is VirtualButton btn) {
-                        taken += Input.GuiButton(btn).Width * 0.6f;
-                    }
-                }
-
-                Vector2 pos = position + new Vector2(Container.Width - taken, 0f) / 2f;
-                for (int i = 0; i < info.Count; i++) {
-                    if (info[i] is string text) {
-                        ActiveFont.DrawOutline(text, pos, new Vector2(0f, 0.5f), Vector2.One * 0.6f, textColor, 2f, strokeColor);
-                        pos.X += ActiveFont.Measure(text).X * 0.6f;
-
-                    } else if (info[i] is VirtualButton btn) {
-                        MTexture tex = Input.GuiButton(btn);
-                        tex.DrawJustified(pos, new Vector2(0f, 0.5f), btnColor, 0.6f);
-                        pos.X += tex.Width * 0.6f;
-                    }
-                }
             }
 
         }
