+++ Celeste.Mod.mm/Patches/Celeste.cs
@@ -243,11 +243,6 @@ https://discord.gg/6qjaePQ");
             // DON'T! The original method is orig_LoadContent
             bool firstLoad = this.firstLoad;
 
-            if (!firstLoad && Version <= new Version(1, 3, 1, 2)) {
-                // Celeste 1.3.1.2 and older runs Directory.Add in PlaybackData.Load
-                PlaybackData.Tutorials.Clear();
-            }
-
             orig_LoadContent();
 
             foreach (EverestModule mod in Everest._Modules)
