+++ Celeste.Mod.mm/Patches/Monocle/VirtualAxis.cs
@@ -4,8 +4,6 @@ using System.Collections.Generic;
 
 namespace Monocle {
     // Copied from Celeste 1.3.3.11
-    [MonoModIfFlag("V2:Input")]
-    // Original name because this type is missing from 1.3.3.14+, thus MonoModPatch isn't properly respected.
     public class VirtualAxis : VirtualInput {
 
         public List<Node> Nodes;
@@ -44,12 +42,10 @@ namespace Monocle {
 
 
 
-        [MonoModIfFlag("V2:Input")]
         public abstract class Node : VirtualInputNode {
             public abstract float Value { get; }
         }
 
-        [MonoModIfFlag("V2:Input")]
         public class PadLeftStickX : Node {
             public int GamepadIndex;
 
@@ -63,7 +59,6 @@ namespace Monocle {
             }
         }
 
-        [MonoModIfFlag("V2:Input")]
         public class PadLeftStickY : Node {
             public int GamepadIndex;
 
@@ -77,7 +72,6 @@ namespace Monocle {
             }
         }
 
-        [MonoModIfFlag("V2:Input")]
         public class PadRightStickX : Node {
             public int GamepadIndex;
 
@@ -91,7 +85,6 @@ namespace Monocle {
             }
         }
 
-        [MonoModIfFlag("V2:Input")]
         public class PadRightStickY : Node {
             public int GamepadIndex;
 
@@ -105,7 +98,6 @@ namespace Monocle {
             }
         }
 
-        [MonoModIfFlag("V2:Input")]
         public class PadDpadLeftRight : Node {
             public int GamepadIndex;
 
@@ -126,7 +118,6 @@ namespace Monocle {
             }
         }
 
-        [MonoModIfFlag("V2:Input")]
         public class PadDpadUpDown : Node {
             public int GamepadIndex;
 
@@ -147,7 +138,6 @@ namespace Monocle {
             }
         }
 
-        [MonoModIfFlag("V2:Input")]
         public class KeyboardKeys : Node {
             public OverlapBehaviors OverlapBehavior;
